PDI-1261 -- Usage of the constructor BigDecimal(double) is discouraged because it produces unpredictable results.  This corrects an error in Kettle's ROUND function of the Calculator step
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@7905 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src-core/org/pentaho/di/compatibility/ValueInteger.java
@@ -149,7 +149,7 @@ public class ValueInteger implements ValueInterface, Cloneable
 
     public BigDecimal getBigNumber()
     {
-        return new BigDecimal((double)number);
+        return new BigDecimal(number);
     }
 
     public void setBigNumber(BigDecimal number)
